requirements.txt
# 保留UniAD必需且无冲突的依赖（删除了mmcv/mmdet/numpy/torch等冲突项）
absl-py==2.3.1
addict==2.4.0
aidisdk==1.1.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backports-zoneinfo==0.2.1
beautifulsoup4==4.14.2
black==24.8.0
bleach==6.1.0
blinker==1.8.2
boto3==1.26.133
botocore==1.29.133
casadi==3.6.7
certifi==2025.10.5
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
configargparse==1.7.1
contourpy==1.1.1
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
dash==3.3.0
dataclass-type-validator==0.0.7
dataclasses-json==0.5.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
dictor==0.1.11
dill==0.3.6
distro==1.9.0
easydict==1.10
einops==0.8.1
entrypoints==0.4
ephem==4.1.4
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.14.0
fire==0.7.1
flake8==7.1.2
flask==3.0.3
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.0
future==1.0.0
google-auth==2.42.1
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==2.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.1
jmespath==0.10.0
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.10.0
jupyter-lsp==2.3.0
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.8
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
keras==2.15.0
kiwisolver==1.4.7
llvmlite==0.31.0
lmdb==1.4.0
lyft-dataset-sdk==0.0.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
model-index==0.1.11
motmetrics==1.1.3
mypy-extensions==1.1.0
narwhals==1.42.1
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.5
notebook==7.3.3
notebook-shim==0.2.4
numba==0.48.0
nuscenes-devkit==1.1.9
oauthlib==3.3.1
# open3d==0.19.0
openai==2.2.0
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
# openssl==3.0.18
# openxlab==0.1.3
openxlab
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==6.4.0
pluggy==1.5.0
plyfile==1.0.3
polars==1.8.2
prettytable==3.11.0
prometheus-client==0.21.1
prompt-toolkit==3.0.52
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
py-trees==0.8.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.10.6
pydantic-core==2.27.2
pydot==1.4.2
pyflakes==3.2.0
pygame==2.3.0
pygments==2.19.2
pyparsing==3.1.4
pypika==0.48.9
pyquaternion==0.9.9
pyrsistent==0.19.3
pytest==8.3.5
pytest-runner==6.0.0
python-json-logger==4.0.0
pytorch-lightning==1.2.5
pytz==2023.4
pywavelets==1.4.1
pyyaml==6.0.3
pyzmq==27.1.0
readline
referencing==0.35.1
# requests==2.25.1
requests==2.31
requests-oauthlib==2.0.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.20.1
rsa==4.9.1
s3transfer==0.6.2
safetensors==0.5.3
send2trash==1.8.3
similaritymeasures==0.7.0
simple-watchdog-timer==0.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
# sqlite==3.50.2
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
threaded==4.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.22
tinycss2==1.2.1
# tk==8.6.15
tomli==2.3.0
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
trimesh==2.35.39
typed-ast==1.5.4
typeguard==2.13.3
typing-extensions==4.13.2
typing-inspect==0.8.0
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.14
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
wheel==0.44.0
widgetsnbextension==4.0.15
# xorg-libx11==1.8.12
# xorg-libxau==1.0.12
# xorg-libxdmcp==1.1.5
# xorg-xorgproto==2024.1
# xz==5.6.4
yapf==0.40.1
# zipp==3.20.2
# zlib==1.3.1